+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KZ1FPKRCD0WAH8MXTYESCAGM"
+        "01KZ40AYH8PBPD3M0H93VT2KVQ"
     )
     r"""The unique identifier of the data source"""