+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -250,9 +250,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-08-02T14:57:47.254Z"
+    created: Optional[str] = "2026-08-03T14:26:59.722Z"
 
-    updated: Optional[str] = "2026-08-02T14:57:47.254Z"
+    updated: Optional[str] = "2026-08-03T14:26:59.722Z"
 
     updated_by_id: OptionalNullable[str] = UNSET