+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -247,7 +247,7 @@ class ListDatasourcesData(BaseModel):
     metadata: ListDatasourcesMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KZ1FPKRAH25MW7D20CVMX8S0"
+        "01KZ40AYH56KW5ERYGHV6Z96HN"
     )
     r"""The unique identifier of the data source"""