+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -811,7 +811,7 @@ class RetrieveDatapointEvaluations4(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-08-02T14:57:57.571Z")
+    ] = parse_datetime("2026-08-03T14:27:10.039Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -960,7 +960,7 @@ class RetrieveDatapointEvaluations3(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-08-02T14:57:57.565Z")
+    ] = parse_datetime("2026-08-03T14:27:10.038Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1105,7 +1105,7 @@ class RetrieveDatapointEvaluations2(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-08-02T14:57:57.564Z")
+    ] = parse_datetime("2026-08-03T14:27:10.038Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1250,7 +1250,7 @@ class RetrieveDatapointEvaluations1(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-08-02T14:57:57.564Z")
+    ] = parse_datetime("2026-08-03T14:27:10.037Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1365,7 +1365,7 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-08-02T14:57:45.633Z")
+    updated: Optional[datetime] = parse_datetime("2026-08-03T14:26:57.829Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")