+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KZ1FPKRCWNB8WDWJKZ414GDV"
+        "01KZ40AYH7X9KVN917F6K4DFQE"
     )
     r"""The unique identifier of the data source"""