+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KZ1FPKREY8KY2FBJGKWD7463"
+        "01KZ40AYHEREMES3KE06454QY5"
     )
     r"""The unique identifier of the data source"""