+++ packages/orq-rc/RELEASES.md
@@ -9929,3 +9929,13 @@ Based on:
 - [python v4.13.0-rc.26] packages/orq-rc
 ### Releases
 - [PyPI v4.13.0-rc.26] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.26 - packages/orq-rc
+
+## 2026-08-03 14:30:22
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.1 (2.926.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.13.0-rc.29] packages/orq-rc
+### Releases
+- [PyPI v4.13.0-rc.29] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.29 - packages/orq-rc